+++ api/src/interfaces.ts
@@ -1,0 +1,9 @@
+export interface productData {
+  id: String;
+  appId: String;
+  inSubscription: Boolean;
+  title: String;
+  subtitle: String;
+  description: String;
+  subscription: String;
+}
